requirements.txt
Flask
flask_sqlalchemy
flask_migrate
python-dotenv
pymysql
requests
pillow
matplotlib